+++ php/tests/compile_extension.sh
@@ -1,20 +1,26 @@
 #!/bin/bash
 
-set -ex
+set -e
 
 cd $(dirname $0)
 
 ../prepare_c_extension.sh
-pushd  ../ext/google/protobuf
-phpize --clean
-rm -f configure.in configure.ac
-phpize
-if [ "$1" = "--release" ]; then
-  ./configure --with-php-config=$(which php-config)
-else
-  # To get debugging symbols in PHP itself, build PHP with:
-  #   $ ./configure --enable-debug CFLAGS='-g -O0'
-  ./configure --with-php-config=$(which php-config) CFLAGS="-g -O0 -Wall"
+pushd  ../ext/google/protobuf > /dev/null
+
+CONFIGURE_OPTIONS=("./configure" "--with-php-config=$(which php-config)")
+
+if [ "$1" != "--release" ]; then
+  CONFIGURE_OPTIONS+=("CFLAGS=-g -O0 -Wall")
 fi
+
+# If the PHP interpreter we are building against or the arguments
+# have changed, we must regenerated the Makefile.
+if [[ ! -f Makefile ]] || [[ "$(grep '  \$ ./configure' config.log)" != "  $ ${CONFIGURE_OPTIONS[@]}" ]]; then
+  phpize --clean
+  rm -f configure.in configure.ac
+  phpize
+  "${CONFIGURE_OPTIONS[@]}"
+fi
+
 make
-popd
+popd > /dev/null
